requirements.txt
Flask==1.1.2
flaskapp==1.0.5
json5==0.9.0
jsonschema==3.2.0
line-bot-sdk==1.17.0
gunicorn==20.0.4
requests==2.22.0
requests-oauthlib==1.3.0
python-dotenv==0.15.0